moved interaces to utils folder
using interfaces to check response data, decided to move to utils as anyone could need them for any object
