+++ src/components/Wiki/Landing/Landing.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {Card, ListGroupItem} from 'react-bootstrap';
+import {Card} from 'react-bootstrap';
 
 import './Landing.css';
