+++ src/utils/responseInterfaces/commentData.ts
@@ -1,0 +1,7 @@
+export interface commentData{
+    id: number,
+    text: string,
+    dateTime: string,
+    user: number,
+    replies: number[]
+}
