+++ src/utils/responseInterfaces/entryData.ts
@@ -1,0 +1,10 @@
+export interface entryData{
+    id: number
+    title: string
+    text: string
+    sideBar: number
+    comments: number[]
+    contributors: number[]
+    headings: number[]
+    log: number[]
+}
