+++ src/utils/responseInterfaces/wikiData.ts
@@ -1,0 +1,6 @@
+export interface wikiData {
+    id: number;
+    title: string;
+    briefDescription: string;
+    entries: number[];
+}
